### 2025-11-10 #9389 @editorialbot
[PRE REVIEW]: JSOSolvers.jl: Unconstrained and bound-constrained optimization solvers
**Submitting author:** @tmigot (<a href="http://orcid.org/0000-0001-7729-2513">Tangi Migot</a>)
**Repository:** https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl
**Branch with paper.md** (empty if default branch): joss-paper-branch
**Version:** 0.14.3
