### 2025-11-10 7b559cf
try to fix DOI
