+++ paper/paper.bib
@@ -18,7 +18,8 @@
   volume={236},
   pages={110651},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.2139/ssrn.4601442}
 }
 
 # Citation for R2, FOMO is an extension.
@@ -30,7 +31,8 @@
   number={2},
   pages={900--929},
   year={2022},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/21m1409536}
 }
 
 @software{Percival_jl,
@@ -40,7 +42,8 @@ month = jul,
 title = {{{Percival.jl: an augmented Lagrangian method}}},
 url = {https://github.com/JuliaSmoothOptimizers/Percival.jl},
 version = {0.7.0},
-year = {2023}
+year = {2023},
+doi = {10.5281/zenodo.3969045}
 }
 
 @software{FletcherPenaltySolver_jl,
@@ -50,7 +53,8 @@ month = jul,
 title = {{{FletcherPenaltySolver.jl: Fletcher's penalty method for nonlinear optimization models}}},
 url = {https://github.com/JuliaSmoothOptimizers/FletcherPenaltySolver.jl},
 version = {0.2.4},
-year = {2023}
+year = {2023},
+doi = {10.5281/zenodo.7153564}
 }
 
 @software{OptimizationProblems_jl,
@@ -60,7 +64,8 @@ month = jul,
 title = {{OptimizationProblems.jl: A collection of optimization problems in Julia}},
 url = {https://github.com/JuliaSmoothOptimizers/OptimizationProblems.jl},
 version = {0.7.3},
-year = {2023}
+year = {2023},
+doi = {10.5281/zenodo.3672094}
 }
 
 @software{The_JuliaSmoothOptimizers_Ecosystem,
@@ -80,7 +85,8 @@ month = mar,
 title = {{{NLPModels.jl: Data Structures for Optimization Models}}},
 url = {https://github.com/JuliaSmoothOptimizers/NLPModels.jl},
 version = {0.20.0},
-year = {2023}
+year = {2023},
+doi = {10.5281/zenodo.2558627}
 }
 
 @article{jump,
@@ -207,7 +213,8 @@ author = {Dussault, Jean-Pierre and Goyette, Samuel and Migot, Tangi and Orban, 
 license = {MIT},
 title = {{{AdaptiveRegularization.jl: A unified efficient implementation of trust-region type algorithms for unconstrained optimization}}},
 url = {https://github.com/JuliaSmoothOptimizers/AdaptiveRegularization.jl},
-year = {2024}
+year = {2024},
+doi = {10.5281/zenodo.10434673}
 }
 
 @software{SolverBenchmark_jl,
@@ -217,7 +224,8 @@ month = nov,
 title = {{{SolverBenchmark.jl: Benchmark Tools for Solvers}}},
 url = {https://github.com/JuliaSmoothOptimizers/SolverBenchmark.jl},
 version = {0.6.0},
-year = {2023}
+year = {2023},
+doi = {10.5281/zenodo.4154524}
 }
 
 @article{dolan2002benchmarking,
@@ -254,7 +262,7 @@ year = {2023}
   pages={35--59},
   year={2006},
   publisher={Springer},
-  doi={10.1007/0-387-30065-1\_4}
+  doi={10.1007/0-387-30065-1_4}
 }
 
 @article{wachter2006implementation,
