### 2025-11-11 52b231a
Add line search params to R2N and refactor tests
Introduces configurable line search parameters (ls_c, ls_increase, ls_decrease, ls_min_alpha, ls_max_alpha) to the R2N solver and parameter set. Refactors the Armijo negative curvature handling to use these parameters. Cleans up and expands test scripts, removes special handling for R2N_exact, and adds a placeholder for HSL subsolver tests.
