+++ my_R2N_dev/test_R2N.jl
@@ -3,7 +3,7 @@ using JSOSolvers
 using HSL
 using NLPModels, ADNLPModels
 using SparseArrays, LinearAlgebra
-using ADNLPModels, Krylov, LinearOperators, NLPModels, NLPModelsModifiers, SolverCore, SolverTools
+using ADNLPModels, Krylov, LinearOperators, NLPModels, NLPModelsModifiers, SolverCore 
 
 # f(x) = (x[1] - 1)^2 + 4 * (x[2] - x[1]^2)^2
 # nlp = ADNLPModel(f, [-1.2; 1.0])
@@ -11,34 +11,58 @@ using ADNLPModels, Krylov, LinearOperators, NLPModels, NLPModelsModifiers, Solve
 #   println("The status is: ", stats.status == :first_order)
 #   println("The solution is approximately: ", isapprox(stats.solution, [1.0; 1.0], atol = 1e-6))
 
-# s = :R2NSolver
-# fun = :R2N
-# f(x) = (x[1] - 1)^2 + 4 * (x[2] - x[1]^2)^2
-# nlp = ADNLPModel(f, [-1.2; 1.0])
-# # nlp.meta.x0 .= 2.0
 
 
-# if fun == :R2N_exact
-#     nlp = LBFGSModel(nlp)
-#     solver = eval(s)(nlp,subsolver= :shifted_lbfgs)
-#   else 
-#     solver = eval(s)(nlp)
-#   end
+# for s in (:R2NSolver,  :FoSolver, :TronSolver, :TrunkSolver)
+for s in (:R2NSolver,  )
+    println("\n\n\t\t===================================")
+    println("============        Testing solver: $s            ==============\n\n")
+    f(x) = (x[1] - 1)^2 + 4 * (x[2] - x[1]^2)^2
+    nlp = ADNLPModel(f, [-1.2; 1.0])
+    solver = eval(s)(nlp)
+    
 
-#   stats = GenericExecutionStats(nlp)
-#   stats = SolverCore.solve!(solver, nlp, stats, verbose = 1)
-#   println("The status is: ", stats.status == :first_order)
-#   println("The solution is approximately: ", isapprox(stats.solution, [1.0; 1.0], atol = 1e-6))
+    stats = GenericExecutionStats(nlp)
+    stats = SolverCore.solve!(solver, nlp, stats, verbose = 1)
+    println("The status is: ", stats.status == :first_order)
+    println("The solution is approximately: ", isapprox(stats.solution, [1.0; 1.0], atol = 1e-6))
 
-#   nlp.meta.x0 .= 2.0
-#   SolverCore.reset!(solver)
+    f2(x) = (x[1])^2 + 4 * (x[2] - x[1]^2)^2
+    nlp = ADNLPModel(f2, [-1.2; 1.0])
+  
+    solver = eval(s)(nlp,subsolver= :cg) 
+    
+    SolverCore.reset!(solver, nlp)
 
-#     stats = GenericExecutionStats(nlp)
-#   stats = SolverCore.solve!(solver, nlp, stats, verbose = 1)
-#   println("The status is: ", stats.status == :first_order)
-#   println("The solution is approximately: ", isapprox(stats.solution, [1.0; 1.0], atol = 1e-6))
+    stats = GenericExecutionStats(nlp)
+    stats = SolverCore.solve!(solver, nlp, stats, verbose = 50,  atol = 1e-10, rtol = 1e-10)
+    println("The status is: ", stats.status == :first_order)
+    println("The solution is approximately: ", isapprox(stats.solution, [0.0; 0.0], atol = 1e-6))
+    print(stats.solution)
+
+end
 
 
[…]
