+++ src/R2N.jl
@@ -73,6 +73,11 @@ struct R2NParameterSet{T} <: AbstractParameterSet
   δ1::Parameter{T, RealInterval{T}}
   σmin::Parameter{T, RealInterval{T}}
   non_mono_size::Parameter{Int, IntegerRange{Int}}
+  ls_c::Parameter{T, RealInterval{T}}
+  ls_increase::Parameter{T, RealInterval{T}}
+  ls_decrease::Parameter{T, RealInterval{T}}
+  ls_min_alpha::Parameter{T, RealInterval{T}}
+  ls_max_alpha::Parameter{T, RealInterval{T}}
 end
 
 # Default parameter values
@@ -92,6 +97,13 @@ const R2N_σmin = DefaultParameter(nlp -> begin
   T(eps(T))
 end, "eps(T)")
 const R2N_non_mono_size = DefaultParameter(1)
+# Line search parameters
+const R2N_ls_c = DefaultParameter(nlp -> eltype(nlp.meta.x0)(1e-4), "T(1e-4)")
+const R2N_ls_increase = DefaultParameter(nlp -> eltype(nlp.meta.x0)(1.5), "T(1.5)")
+const R2N_ls_decrease = DefaultParameter(nlp -> eltype(nlp.meta.x0)(0.5), "T(0.5)")
+const R2N_ls_min_alpha = DefaultParameter(nlp -> eltype(nlp.meta.x0)(1e-8), "T(1e-8)")
+const R2N_ls_max_alpha = DefaultParameter(nlp -> eltype(nlp.meta.x0)(1e2), "T(1e2)")
+
 
 function R2NParameterSet(
   nlp::AbstractNLPModel;
@@ -105,6 +117,11 @@ function R2NParameterSet(
   δ1::T = get(R2N_δ1, nlp),
   σmin::T = get(R2N_σmin, nlp),
   non_mono_size::Int = get(R2N_non_mono_size, nlp),
+  ls_c::T = get(R2N_ls_c, nlp),
+  ls_increase::T = get(R2N_ls_increase, nlp),
+  ls_decrease::T = get(R2N_ls_decrease, nlp),
+  ls_min_alpha::T = get(R2N_ls_min_alpha, nlp),
+  ls_max_alpha::T = get(R2N_ls_max_alpha, nlp),
 ) where {T}
   R2NParameterSet{T}(
     Parameter(θ1, RealInterval(zero(T), one(T))),
@@ -117,6 +134,13 @@ function R2NParameterSet(
     Parameter(δ1, RealInterval(zero(T), one(T))),
     Parameter(σmin, RealInterval(zero(T), T(Inf))),
     Parameter(non_mono_size, IntegerRange(1, typemax(Int))),
+
+    Parameter(ls_c, RealInterval(zero(T), one(T))), # c is typically (0, 1)
+    Parameter(ls_increase, RealInterval(one(T), T(Inf))), # increase > 1
+    Parameter(ls_decrease, RealInterval(zero(T), one(T))), # decrease < 1
+    Parameter(ls_min_alpha, RealInterval(zero(T), T(Inf))),
+    Parameter(ls_max_alpha, RealInterval(zero(T), T(Inf))),
+    
   )
 end
 
@@ -216,7 +240,7 @@ For advanced usage, first define a `R2NSolver` to preallocate the memory used in
 - `max_time::Float64 = 30.0`: maximum time limit in seconds.
 - `max_iter::Int = typemax(Int)`: maximum number of iterations.
 - `verbose::Int = 0`: if > 0, display iteration details every `verbose` iteration.
[…]

[…]
