### 2025-11-14 3b66093
Enhance HSL direct solver integration and test coverage
Refactored and improved the integration of HSL direct solvers (MA97, MA57) in R2N, including more robust handling, logging, and test coverage. Enabled and updated tests for HSL solvers, improved output verbosity, and fixed minor issues in test and solver logic. Also adjusted test verbosity and temporarily disabled a problematic test include.
