+++ src/R2N.jl
@@ -6,14 +6,10 @@ using LinearOperators, LinearAlgebra
 using SparseArrays
 using HSL
 
-#TODO opnorm (f norm) of coo matrix
-# then move it to SparseMatrixCOO
-# Keep H as Sparse
-
-
 #TODO move to LinearOperators
 # Define a new mutable operator for A = H + σI
-mutable struct ShiftedOperator{T, V, OpH <: Union{AbstractLinearOperator{T}, AbstractMatrix{T}}} <: AbstractLinearOperator{T}
+mutable struct ShiftedOperator{T, V, OpH <: Union{AbstractLinearOperator{T}, AbstractMatrix{T}}} <:
+               AbstractLinearOperator{T}
   H::OpH
   σ::T
   n::Int
@@ -21,7 +17,9 @@ mutable struct ShiftedOperator{T, V, OpH <: Union{AbstractLinearOperator{T}, Abs
   hermitian::Bool
 end
 
-function ShiftedOperator(H::OpH) where {T, OpH <: Union{AbstractLinearOperator{T}, AbstractMatrix{T}}}
+function ShiftedOperator(
+  H::OpH,
+) where {T, OpH <: Union{AbstractLinearOperator{T}, AbstractMatrix{T}}}
   is_sym = isa(H, AbstractLinearOperator) ? H.symmetric : issymmetric(H)
   is_herm = isa(H, AbstractLinearOperator) ? H.hermitian : ishermitian(H)
   return ShiftedOperator{T, Vector{T}, OpH}(H, zero(T), size(H, 1), is_sym, is_herm)
@@ -43,7 +41,6 @@ function LinearAlgebra.mul!(
   LinearAlgebra.axpy!(A.σ, x, y)
   return y
 end
-
 
 """
     R2NParameterSet([T=Float64]; θ1, θ2, η1, η2, γ1, γ2, γ3, σmin, non_mono_size)
@@ -104,7 +101,6 @@ const R2N_ls_decrease = DefaultParameter(nlp -> eltype(nlp.meta.x0)(0.5), "T(0.5
 const R2N_ls_min_alpha = DefaultParameter(nlp -> eltype(nlp.meta.x0)(1e-8), "T(1e-8)")
 const R2N_ls_max_alpha = DefaultParameter(nlp -> eltype(nlp.meta.x0)(1e2), "T(1e2)")
 
-
 function R2NParameterSet(
   nlp::AbstractNLPModel;
   θ1::T = get(R2N_θ1, nlp),
@@ -134,13 +130,11 @@ function R2NParameterSet(
     Parameter(δ1, RealInterval(zero(T), one(T))),
     Parameter(σmin, RealInterval(zero(T), T(Inf))),
     Parameter(non_mono_size, IntegerRange(1, typemax(Int))),
-
     Parameter(ls_c, RealInterval(zero(T), one(T))), # c is typically (0, 1)
     Parameter(ls_increase, RealInterval(one(T), T(Inf))), # increase > 1
     Parameter(ls_decrease, RealInterval(zero(T), one(T))), # decrease < 1
     Parameter(ls_min_alpha, RealInterval(zero(T), T(Inf))),
     Parameter(ls_max_alpha, RealInterval(zero(T), T(Inf))),
-    
   )
 end
 
@@ -160,13 +154,13 @@ HSLDirectSolver: Generic wrapper for HSL direct solvers (e.g., MA97, MA57).
 - nnzh: number of Hessian nonzeros
 """
 mutable struct HSLDirectSolver{T, S} <: AbstractMASolver
-    hsl_obj::S
-    rows::Vector{Int}
-    cols::Vector{Int}
-    vals::Vector{T}
-    n::Int
[…]
