+++ my_R2N_dev/test_R2N.jl
@@ -52,17 +52,68 @@ nlp= ADNLPModel(
     )
 
 stats_armjio = R2N(nlp, verbose = 1, max_iter=5000, subsolver= :minres_qlp, npc_handler= :armijo)
+stats_hsl= R2N(nlp, verbose = 10, max_iter=5000, subsolver= :ma97) # todo we need this to handle npc
 stats_sigma = R2N(nlp, verbose = 10, max_iter=5000, subsolver= :minres_qlp, npc_handler= :sigma)
 stats_prev = R2N(nlp, verbose = 10, max_iter=5000, subsolver= :minres_qlp, npc_handler= :prev)
 stats_cp = R2N(nlp, verbose = 10, max_iter=5000, subsolver= :minres_qlp, npc_handler= :cp)  
-stats_trunk = trunk(nlp, verbose = 10, max_iter=5000)
+stats_trunk = trunk(nlp, verbose = 1, max_iter=5000)
 println("The stats after 5k iteration is")
-println("Armijo: ", stats_armjio.status, " max_iter :" , stats_armjio.iter)
-println("Sigma: ", stats_sigma.status, " max_iter :" , stats_sigma.iter)
-println("Previous: ", stats_prev.status, " max_iter :" , stats_prev.iter)
-println("Cauchy Point: ", stats_cp.status, " max_iter :" , stats_cp.iter)
-println("Truncated: ", stats_trunk.status, " max_iter :" , stats_trunk.iter)
+println("Armijo: ", stats_armjio.status, " max_iter :" , stats_armjio.iter, " solution is ", stats_armjio.solution)
+println("Sigma: ", stats_sigma.status, " max_iter :" , stats_sigma.iter, " solution is ")#, stats_sigma.solution)
+println("Previous: ", stats_prev.status, " max_iter :" , stats_prev.iter, " solution is ")#, stats_prev.solution)
+println("Cauchy Point: ", stats_cp.status, " max_iter :" , stats_cp.iter, " solution is ")#, stats_cp.solution)
+println("trunk: ", stats_trunk.status, " max_iter :" , stats_trunk.iter, " solution is ", stats_trunk.solution)
+println("HSL: ", stats_hsl.status, " max_iter :" , stats_hsl.iter, " solution is ", stats_hsl.solution)
 println("\n\n\t\t===================================")
+
+
+
+
+
+
+
+# multiple precisions
+T = Float32
+f(x) = (x[1] - 1)^2 + 4 * (x[2] - x[1]^2)^2
+x0 = T[-1.2; 1.0]
+nlp = ADNLPModel(f, x0)
+
+
+solver = R2N
+# solver = (nlp; kwargs...) -> R2N(LBFGSModel(nlp), subsolver= :shifted_lbfgs; kwargs...)
+
+ϵ = eps(T)^T(1 / 4)
+
+ng0 = norm(grad(nlp, nlp.meta.x0))
+
+stats = solver(nlp, verbose = 1, subsolver_verbose = 1 , atol = ϵ, rtol = ϵ, subsolver= :cr)
+
+
+print(" eltype(stats.solution):", eltype(stats.solution))
+print( "stats.dual_feas ", stats.dual_feas, " ϵ * ng0 + ϵ: ", ϵ * ng0 + ϵ)
+print( "=====================================")
+
+
+##### R2N and lbfgs exact solver
+
+println("R2N with LBFGS exact solver as subsolver")
+n = 30
+nlp= ADNLPModel(
+      x -> sum(100 * (x[i + 1] - x[i]^2)^2 + (x[i] - 1)^2 for i = 1:(n - 1)),
+      collect(1:n) ./ (n + 1),
+      name = "Extended Rosenbrock";
+    )
[…]
