### 2025-11-19 #9389 @nkrusch
Hello @dourouc05, @raphaelchenouard, and @GuillaumeDerval,

I am inviting you to review this submission for JOSS.
Please, let me know if you are interested.
