### 2025-11-19 ba51e77
Update LICENSE.md to reflect MPL version 2.0
Removed copyright notice and license introduction.
