+++ LICENSE.md
@@ -1,9 +1,3 @@
-Copyright (c) 2019-present: Abel Soares Siqueira and Dominique Orban
-
-JSOSolvers.jl is licensed under the [MPL version 2.0](https://www.mozilla.org/MPL/2.0/).
-
-## License
-
 	Mozilla Public License Version 2.0
 	==================================
 
