### 2025-11-19 71e4094
Format LICENSE.md for consistency and readability
Formatted the text for better readability and consistency.
