+++ LICENSE.md
@@ -1,373 +1,373 @@
-	Mozilla Public License Version 2.0
-	==================================
-
-	1. Definitions
-	--------------
-
-	1.1. "Contributor"
-		means each individual or legal entity that creates, contributes to
-		the creation of, or owns Covered Software.
-
-	1.2. "Contributor Version"
-		means the combination of the Contributions of others (if any) used
-		by a Contributor and that particular Contributor's Contribution.
-
-	1.3. "Contribution"
-		means Covered Software of a particular Contributor.
-
-	1.4. "Covered Software"
-		means Source Code Form to which the initial Contributor has attached
-		the notice in Exhibit A, the Executable Form of such Source Code
-		Form, and Modifications of such Source Code Form, in each case
-		including portions thereof.
-
-	1.5. "Incompatible With Secondary Licenses"
-		means
-
-		(a) that the initial Contributor has attached the notice described
-			in Exhibit B to the Covered Software; or
+Mozilla Public License Version 2.0
+==================================
+
+1. Definitions
+--------------
+
+1.1. "Contributor"
+    means each individual or legal entity that creates, contributes to
+    the creation of, or owns Covered Software.
+
+1.2. "Contributor Version"
+    means the combination of the Contributions of others (if any) used
+    by a Contributor and that particular Contributor's Contribution.
+
+1.3. "Contribution"
+    means Covered Software of a particular Contributor.
+
+1.4. "Covered Software"
+    means Source Code Form to which the initial Contributor has attached
+    the notice in Exhibit A, the Executable Form of such Source Code
+    Form, and Modifications of such Source Code Form, in each case
+    including portions thereof.
+
+1.5. "Incompatible With Secondary Licenses"
+    means
+
+    (a) that the initial Contributor has attached the notice described
+        in Exhibit B to the Covered Software; or
 
-		(b) that the Covered Software was made available under the terms of
-			version 1.1 or earlier of the License, but not also under the
-			terms of a Secondary License.
+    (b) that the Covered Software was made available under the terms of
+        version 1.1 or earlier of the License, but not also under the
+        terms of a Secondary License.
 
-	1.6. "Executable Form"
-		means any form of the work other than Source Code Form.
+1.6. "Executable Form"
+    means any form of the work other than Source Code Form.
 
-	1.7. "Larger Work"
-		means a work that combines Covered Software with other material, in
-		a separate file or files, that is not Covered Software.
+1.7. "Larger Work"
+    means a work that combines Covered Software with other material, in
+    a separate file or files, that is not Covered Software.
 
-	1.8. "License"
-		means this document.
+1.8. "License"
+    means this document.
 
-	1.9. "Licensable"
-		means having the right to grant, to the maximum extent possible,
-		whether at the time of the initial grant or subsequently, any and
-		all of the rights conveyed by this License.
+1.9. "Licensable"
+    means having the right to grant, to the maximum extent possible,
+    whether at the time of the initial grant or subsequently, any and
+    all of the rights conveyed by this License.
 
-	1.10. "Modifications"
-		means any of the following:
+1.10. "Modifications"
+    means any of the following:
 
-		(a) any file in Source Code Form that results from an addition to,
-			deletion from, or modification of the contents of Covered
-			Software; or
+    (a) any file in Source Code Form that results from an addition to,
+        deletion from, or modification of the contents of Covered
+        Software; or
 
-		(b) any new file in Source Code Form that contains any Covered
-			Software.
+    (b) any new file in Source Code Form that contains any Covered
+        Software.
 
-	1.11. "Patent Claims" of a Contributor
-		means any patent claim(s), including without limitation, method,
-		process, and apparatus claims, in any patent Licensable by such
[…]
