### 2025-11-19 f2b347a
Change license md (#324)
* Update LICENSE.md to reflect MPL version 2.0
