### 2025-11-19 bf4827e
Change license md (#324)
* Update LICENSE.md to reflect MPL version 2.0
