### 2025-11-22 e27ba53
Clarify phrasing
