### 2025-11-22 e053c22
change figure format
