### 2025-11-22 373bfb8
add list of GPU solvers currently supported
