### 2025-11-22 0453ad4
Add GPU test for TRUNK, TRUNK-NLS and FOMO
