+++ test/runtests.jl
@@ -8,6 +8,8 @@ using NLPModelsTest, SolverParameters
 # this package
 using JSOSolvers
 
+include("test-gpu.jl")
+
 @testset "Test parameterset" begin
   @testset "Test unconstrained parameters $paramset" for (paramset, fun) in (
     (LBFGSParameterSet, lbfgs),
