+++ previews/PR328/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"performance-tips/#Performance-Tips","page":"Performance Tips","title":"Performance Tips","text":"The main functions have an in-place variant that allows one to solve multiple optimization problems with the same dimensions and data types.\n\nMore general solvers can take advantage of this functionality by allocating workspace for the solve only once. The in-place variants only require a Julia structure that contains all the storage needed as an additional argument. In-place methods limit memory allocations and deallocations, which are particularly expensive on GPUs.\n\nusing NLPModels, NLPModelsTest\nnlp = BROWNDEN() # test problem from NLPModelsTest with allocation free evaluations\n\nusing SolverCore\nstats = GenericExecutionStats(nlp) # pre-allocate output structure\nusing JSOSolvers\nsolver = TronSolver(nlp) # pre-allocate workspace\nSolverCore.solve!(solver, nlp, stats) # call tron in place\nJSOSolvers.reset!(solver)\nNLPModels.reset!(nlp) # reset counters\n(@allocated SolverCore.solve!(solver, nlp, stats)) == 0\n","category":"section"},{"location":"solvers/#Solvers","page":"Solvers","title":"Solvers","text":"Solver list\n\nlbfgs\ntron\ntrunk\nR2\nfomo\n\nProblem type Solvers\nUnconstrained NLP lbfgs, tron, trunk, R2, fomo\nUnconstrained NLS trunk, tron\nBound-constrained NLP tron\nBound-constrained NLS tron","category":"section"},{"location":"solvers/#Solver-list","page":"Solvers","title":"Solver list","text":"","category":"section"},{"location":"solvers/#JSOSolvers.lbfgs","page":"Solvers","title":"JSOSolvers.lbfgs","text":"lbfgs(nlp; kwargs...)\n\nAn implementation of a limited memory BFGS line-search method for unconstrained minimization.\n\nFor advanced usage, first define a LBFGSSolver to preallocate the memory used in the algorithm, and then call solve!.\n\nsolver = LBFGSSolver(nlp; mem::Int = 5)\nsolve!(solver, nlp; kwargs...)\n\nArguments\n\nnlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.\n\nThe keyword arguments may include\n\nx::V = nlp.meta.x0: the initial guess.\nmem::Int = 5: algorithm parameter, see LBFGSParameterSet.\natol::T = √eps(T): absolute tolerance.\nrtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.\nmax_eval::Int = -1: maximum number of objective function evaluations.\nmax_time::Float64 = 30.0: maximum time limit in seconds.\nmax_iter::Int = typemax(Int): maximum number of iterations.\nτ₁::T = T(0.9999): algorithm parameter, see LBFGSParameterSet.\nbk_max:: Int = 25: algorithm parameter, see LBFGSParameterSet.\nverbose::Int = 0: if > 0, display iteration details every verbose iteration.\nverbose_subsolver::Int = 0: if > 0, display iteration information every verbose_subsolver iteration of the subsolver.\n\nOutput\n\nThe returned value is a GenericExecutionStats, see SolverCore.jl.\n\nCallback\n\nThe callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should 
[…]
