### 2025-11-22 JuliaSmoothOptimizers/JSOSolvers.jl#329 @giovannifereoli
Documentation link and a few documentation improvement suggestions
Hi! Maybe it’s just me, but I wasn’t able to find the documentation link in the README.

- [ ] It might be helpful to add:

[https://jso.dev/JSOSolvers.jl/stable/](https://jso.dev/JSOSolvers.jl/stable/)

I also had a few small suggestions regarding the documentation:

- [ ] There’s currently no explicit **Statement of Need**. Even a couple of sentences summarizing the purpose and motivation of the package (based on the papers) would help new users understand the scope.
- [ ] The examples in the *Solvers* section are useful, but it might be even better to have a dedicated **Examples** section with a couple of longer scripts - one for bounded and one for unbounded problems, possibly with plots or more complete workflows.
- [ ] Installation is easy (`pkg> add JSOSolvers`), but it could still be helpful to state it explicitly in the docs for completeness.
- [ ] I don’t see any **community guidelines** (e.g., contributing guidelines or a code of conduct).

Thanks again for all the great work!
