### 2025-11-22 #9467 @giovannifereoli
I’ve opened an issue with some documentation suggestions (still working through it, but here it is for now): [https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/issues/329](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/issues/329)
