### 2025-11-22 #9467 @giovannifereoli
While running `benchmarks.jl` and analyzing the code, I noticed that several packages were missing on my side (`NLPModelsIpopt`, `SolverBenchmark`, `CUTEst`, `DataFrames`, `Dates`, `JLD2`). That part was on me, so no issue there.

However, there does seem to be a small bug in the benchmark script: at line 18 it throws an error about a missing type parameter `T` for the expression

```julia
tol = sqrt(eps(T))  # relative tolerance
```

It looks like the function is expecting a type `T` but never defines it. Maybe those scripts used for the benchmarks could be cleaned up and adjusted a bit. It would be great to be able to run the full benchmark suite at least once!
