### 2025-11-23 f4984d6
Apply comments to the paper (#327)
* Clarify phrasing

* change figure format

* add list of GPU solvers currently supported
