+++ paper/paper.md
@@ -56,9 +56,9 @@ A nonlinear least-squares problem is a special case of \eqref{eq:nlp}, where $f(
 TRON and TRUNK have specialized implementations leveraging the structure of residual models to improve performance and scalability.
 
 A key strength of `JSOSolvers.jl` lies in its efficiency and flexibility.
-The solvers support fully in-place execution, allowing repeated solves with zero memory allocation, which is particularly beneficial in high-performance and GPU computing environments where memory management is critical.
+The solvers support fully in-place execution, allowing repeated solves without additional memory allocation, which is particularly beneficial in high-performance and GPU computing environments where memory management is critical.
 The solvers support any floating-point type, including extended and multi-precision types such as BigFloat, DoubleFloats or QuadMath.
-In addition, several solvers support GPU arrays, broadening the range of hardware where the package can be effectively deployed, for instance when used together with `ExaModels.jl` [@shin2024accelerating].
+Moreover, TRUNK, TRUNK-NLS, and FOMO support GPU arrays, broadening the range of hardware where the package can be effectively deployed, for instance when used together with `ExaModels.jl` [@shin2024accelerating].
 The package documentation and \url{https://jso.dev/tutorials} provide examples illustrating the use of different floating-point systems.
 Furthermore, the solvers expose in-place function variants, allowing multiple optimization problems with identical dimensions and data types to be solved efficiently without reallocations.
 
@@ -107,10 +107,11 @@ include("benchmark.jl") # run the benchmark and store the result in a JLD2 file
 include("analyze_benchmark.jl") # make the figure
 ```
 -->
-![Unconstrained solvers on CUTEst with respect to the elapsed time.](2025-09-06_ipopt_lbfgs_trunk_tron_cutest_Float64_0_291_time_pp.png){ width=100% }
+![Unconstrained solvers on CUTEst with respect to the elapsed time.](2025-09-06_ipopt_lbfgs_trunk_tron_cutest_Float64_0_291_time_pp.pdf){ width=100% }
 
 # Acknowledgements
 
 Dominique Orban is partially supported by an NSERC Discovery Grant.
 
 # References
+
