### 2025-11-23 d1d8693
be more specific in the workflow
