### 2025-11-23 9590323
Replace nightly by pre (#330)
* Replace nightly by pre
