### 2025-11-23 28f3fa1
Add GPU test for TRUNK, TRUNK-NLS and FOMO (#328)
* Add GPU test for TRUNK, TRUNK-NLS and FOMO

* fix typo
