### 2025-11-23 8222c57
Add GPU test for TRUNK, TRUNK-NLS and FOMO (#328)
* Add GPU test for TRUNK, TRUNK-NLS and FOMO

* fix typo
