+++ paper/benchmark.jl
@@ -8,7 +8,7 @@ using SolverBenchmark
 using CUTEst
 
 cutest_problems = readlines("list_problems.dat")
-problems = (CUTEstModel(p) for p in cutest_problems[1:2])
+problems = (CUTEstModel(p) for p in cutest_problems)
 
 max_time = 1200.0 # 20 minutes
 T = Float64
