### 2025-11-24 941910a
Improve HSL direct solver handling and inertia detection
Refactors the main solver loop to better handle singular and indefinite systems when using HSL direct solvers (MA97/MA57), including robust inertia detection and sigma adjustment. Adds get_inertia methods for MA97 and MA57, centralizes HSL factorization/solve logic, and improves non-positive curvature handling. Also adds an example in test_HSL.jl for using MA97 and MA57 with singular systems.
