+++ src/R2N.jl
@@ -599,14 +599,115 @@ function SolverCore.solve!(
 
   ν_k = one(T) # used for scp calculation
   γ_k = zero(T)
-  scp_recal = true
 
   while !done
-    scp_recal = true # Recalculate the Cauchy point if needed
 
+    # Solving for step direction s_k
+    ∇fk .*= -1
+    if r2_subsolver isa CgWorkspace || r2_subsolver isa CrWorkspace
+      # Update the shift in the operator
+      solver.A.σ = σk
+      solver.H = H
+    end
+    subsolver_solved, sub_stats, subiter, npcCount =
+      subsolve!(r2_subsolver, solver, nlp, s, zero(T), n, subsolver_verbose)
+
+    if !subsolver_solved && npcCount == 0
+      @warn("Subsolver failed to solve the system")
+      # TODO exit cleaning 
+      break
+    end
+
+    calc_scp_needed = false
+    force_sigma_increase = false
+
+    if r2_subsolver isa HSLDirectSolver
+      num_neg, num_zero = get_inertia(r2_subsolver)
+      coo_sym_prod!(r2_subsolver.rows, r2_subsolver.cols, r2_subsolver.vals, s, Hs)
+
+      # Check Singularity (Zero Eigenvalues)          # assume Inconsistent could happen then increase the sigma
+      if num_zero > 0
+        force_sigma_increase = true #TODO Prof. Orban, for now we just increase sigma when we have zero eigenvalues
+      end
+
+      # Check Indefinite (Negative Eigenvalues)
+      if !force_sigma_increase && num_neg > 0
+        # Curvature = s' (H+σI) s
+        curv_s = dot(s, Hs)
+
+        if curv_s < 0
+          npcCount = 1 # we need to set this for later use
+          if npc_handler == :prev
+            npc_handler = :armijo  #Force the npc_handler to be armijo and not :prev since we can not have that
+          end
+        else
+          # Step has positive curvature, but matrix has negative eigs.
+          #"compute scp and compare"
+          calc_scp_needed = true
+        end
+      end
+    end
+
+    if !(r2_subsolver isa ShiftedLBFGSSolver) && npcCount >= 1  #npc case
+      if npc_handler == :armijo
+        # --- Algorithm 3: From MINRES Paper --- 
+        npcCount = 0 # TODO check for Subsolver, it need to reset this 
+
+        # 1. Get current state from R2N solver
+        α = one(T) * 1.0     # Initial step size guess
+        f0 = stats.objective   # Current objective value, f(x)
+        if r2_subsolver isa HSLDirectSolver
+          dir = s
+        else
+          dir = r2_subsolver.npc_dir # The non-positive curvature direction
+        end
+        # ∇fk = -∇f(x), so grad = -∇fk
+        grad_dir = dot(-∇fk, dir)
+
+        # 3. First Try (α = 1.0)
+        xt .= x .+ α .* dir
+        fck = obj(nlp, xt) # f(x + α*dir)
+
+        # Check Armijo violation using our new parameter 'ls_c'
+        if fck > f0 + ls_c * α * grad_dir
+          # 4. Backward Tracking (Algorithm 2)
+          while fck > f0 + ls_c * α * grad_dir && α > ls_min_alpha
+            α *= ls_decrease # Use parameter
+            xt .= x .+ α .* dir
+            fck = obj(nlp, xt)
+          end
+        else
+          # 5. Forward Tracking
+          while true
+            α_new = α * ls_increase # Use parameter
[…]
