### 2025-11-25 4cbd7ec
Fix benchmark script and add runner.jl (#331)
* Fix benchmark script and add runner.jl
