### 2025-11-25 7e675ca
Add more information to the documentation
