### 2025-11-25 555d741
Delete 'How to cite' section from index.md
Removed citation section from the documentation.
