### 2025-11-25 f8df78a
Update optimization problem notation in documentation
Replaced LaTeX equation with plain text representation.
