### 2025-11-25 85295f5
Fix formatting in documentation for JSOSolvers
