### 2025-11-25 JuliaSmoothOptimizers/JSOSolvers.jl#329 @giovannifereoli
Not everyone realizes the badge is clickable, so it might help to mention that somewhere - maybe a short note in the README or in the description bar on the upper right?

From a navigation standpoint, I’d also like an *Examples* page in the left-hand docs sidebar (next to Solvers / Internals / Functions). Even a lightweight entry that simply says “Maintaining full examples here is difficult - please refer to the following examples: <link>” would make them much easier to find I believe
