### 2025-11-26 66d9fe6
Add details about encouraging results (#332)
