### 2025-11-26 55f6410
Update test parameters and improve npcCount handling
Commented out unused test code in test_R2N.jl and adjusted test parameters for max_iter and verbose levels. Improved npcCount initialization and reset logic in R2N.jl, and added comments regarding HSL solver flag handling. These changes help clarify test output and address issues with negative curvature count tracking.
