+++ my_R2N_dev/test_R2N.jl
@@ -13,34 +13,34 @@ using ADNLPModels, Krylov, LinearOperators, NLPModels, NLPModelsModifiers, Solve
 
 
 
-# for s in (:R2NSolver,  :FoSolver, :TronSolver, :TrunkSolver)
-for s in (:R2NSolver,  )
-    println("\n\n\t\t===================================")
-    println("============        Testing solver: $s            ==============\n\n")
-    f(x) = (x[1] - 1)^2 + 4 * (x[2] - x[1]^2)^2
-    nlp = ADNLPModel(f, [-1.2; 1.0])
-    solver = eval(s)(nlp)
+# # for s in (:R2NSolver,  :FoSolver, :TronSolver, :TrunkSolver)
+# for s in (:R2NSolver,  )
+#     println("\n\n\t\t===================================")
+#     println("============        Testing solver: $s            ==============\n\n")
+#     f(x) = (x[1] - 1)^2 + 4 * (x[2] - x[1]^2)^2
+#     nlp = ADNLPModel(f, [-1.2; 1.0])
+#     solver = eval(s)(nlp)
     
 
-    stats = GenericExecutionStats(nlp)
-    stats = SolverCore.solve!(solver, nlp, stats, verbose = 1)
-    println("The status is: ", stats.status == :first_order)
-    println("The solution is approximately: ", isapprox(stats.solution, [1.0; 1.0], atol = 1e-6))
+#     stats = GenericExecutionStats(nlp)
+#     stats = SolverCore.solve!(solver, nlp, stats, verbose = 1)
+#     println("The status is: ", stats.status == :first_order)
+#     println("The solution is approximately: ", isapprox(stats.solution, [1.0; 1.0], atol = 1e-6))
 
-    f2(x) = (x[1])^2 + 4 * (x[2] - x[1]^2)^2
-    nlp = ADNLPModel(f2, [-1.2; 1.0])
+#     # f2(x) = (x[1])^2 + 4 * (x[2] - x[1]^2)^2
+#     # nlp = ADNLPModel(f2, [-1.2; 1.0])
   
-    solver = eval(s)(nlp,subsolver= :cg) 
+#     # solver = eval(s)(nlp,subsolver= :cg) 
     
-    SolverCore.reset!(solver, nlp)
+#     # SolverCore.reset!(solver, nlp)
 
-    stats = GenericExecutionStats(nlp)
-    stats = SolverCore.solve!(solver, nlp, stats, verbose = 50,  atol = 1e-10, rtol = 1e-10)
-    println("The status is: ", stats.status == :first_order)
-    println("The solution is approximately: ", isapprox(stats.solution, [0.0; 0.0], atol = 1e-6))
-    print(stats.solution)
+#     # stats = GenericExecutionStats(nlp)
+#     # stats = SolverCore.solve!(solver, nlp, stats, verbose = 50,  atol = 1e-10, rtol = 1e-10)
+#     # println("The status is: ", stats.status == :first_order)
+#     # println("The solution is approximately: ", isapprox(stats.solution, [0.0; 0.0], atol = 1e-6))
+#     # print(stats.solution)
 
-end
+# end
 
 
 println("example of different negative curvature handling strategies with R2N solver")
@@ -51,19 +51,19 @@ nlp= ADNLPModel(
       name = "Extended Rosenbrock";
     )
 
-stats_armjio = R2N(nlp, verbose = 1, max_iter=5000, subsolver= :minres_qlp, npc_handler= :armijo)
-stats_hsl= R2N(nlp, verbose = 10, max_iter=5000, subsolver= :ma97) # todo we need this to handle npc
[…]
