### 2025-11-26 2997bab
Fix heading format in examples.md
