+++ docs/src/examples.md
@@ -1,3 +1,4 @@
-## Examples
+# Examples
 
 Beyond this repository's documentation, you can also find a list of tutorials on [JuliaSmoothOptimizers Tutorials](https://jso.dev/tutorials) by selecting the tag `JSOSolvers.jl`. For instance, the tutorial [Introduction to JSOSolvers](https://jso.dev/tutorials/introduction-to-jsosolvers/#title) is a good starting point.
+
