### 2025-11-26 a465b4a
Update docs/src/index.md
