### 2025-11-26 38c755a
Enhance documentation with function and algorithm details
Added explanation of function properties and optimization methods.
