### 2025-11-27 JuliaSmoothOptimizers/JSOSolvers.jl#329 @giovannifereoli
Heads up - I’m adding an item to the checklist above regarding automated testing!
