### 2025-11-28 JuliaSmoothOptimizers/JSOSolvers.jl#329 @giovannifereoli
I wouldn’t normally think to include those details in the documentation either, but I understand they can be important for contributors. I’m following the JOSS checklist, which asks for this information to be present… Maybe @nkrusch / other reviewers can weigh in on what they prefer?
