### 2025-11-29 aad6f90
Add unit test instructions for JSOSolvers package
Added instructions for running unit tests for the JSOSolvers package.
