### 2025-11-29 3fb70c2
Add unit test instructions to README
Added instructions for running unit tests.
