### 2025-11-29 JuliaSmoothOptimizers/JSOSolvers.jl#329 @tmigot
Thanks for the clarification @nkrusch ! I added instructions to run the unit tests in the doc (which is essentially to do ] test).

For the dependencies, I am indeed a bit unsure what additional information would be useful to add. As is standard in the Julia ecosystem, the package includes a Project.toml file that declares all dependencies and version compatibility constraints. A typical user does not need to install or manage these manually—running Pkg.add("JSOSolvers") and then using JSOSolvers automatically installs the appropriate versions of all required packages.

The JOSS paper already provides some motivation for the key dependencies, but we can certainly clarify this point in the documentation. For example, we could add a short section explaining that dependencies are managed through Project.toml and automatically handled by Julia’s package manager, with no manual action needed from the user.

We want to avoid listing or duplicating all dependencies in the documentation, as this would introduce unnecessary maintenance overhead and risk becoming outdated, but we are happy to add a concise explanation of how dependencies are handled in Julia and by the package.
