### 2025-11-29 815e89b
Add more information to the documentation (#333)
* Add more information to the documentation

* Delete 'How to cite' section from index.md

Removed citation section from the documentation.

* Update optimization problem notation in documentation

Replaced LaTeX equation with plain text representation.

* Fix formatting in documentation for JSOSolvers

* add more

* Apply suggestions from code review

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Fix heading format in examples.md

* Update docs/src/index.md

* Update docs/src/index.md

* Enhance documentation with func
[…]
