### 2025-12-01 2400f5d
Add tests for HSL subsolvers in test suite
Introduces a new test file 'test_hsl_subsolver.jl' to verify the functionality of HSL subsolvers (ma97 and ma57, with and without Armijo handler) using a simple nonlinear problem. The new test is included in 'runtests.jl' and will be skipped if LIBHSL is not functional.
