+++ test/runtests.jl
@@ -90,6 +90,7 @@ end
   end
 end
 
+include("test_hsl_subsolver.jl")
 # include("restart.jl") #TODO issue with rtol -10e-10
 include("callback.jl")
 include("consistency.jl")
