### 2025-12-01 efd86da
Update test_hsl_subsolver.jl
