### 2025-12-05 2dd2165
add default parameters for momentum strategies
