### 2025-12-08 424e0a3
Update docs/src/index.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
