+++ docs/src/index.md
@@ -25,13 +25,45 @@ This strategy enables the solution of large-scale problems even when function an
 pkg> add JSOSolvers
 ```
 
+## Tests
+
 You can run the package’s unit tests with:
 
 ```julia
 pkg> test JSOSolvers
 ```
 
-The tests cover basic solver correctness on analytical test problems, conformity with the NLPModels API, stopping rules and line-search behaviour, and linear algebra components used by second-order methods.
+The test suite is run automatically in continuous integration for every pull request and release, and can also be executed locally as above.
+
+### What the tests cover
+
+The test suite is composed of two layers:
+1. Unit tests and regression tests (solver-specific);
+1. Convergence tests using SolverTest.jl (problem-specific);
+
+#### Unit and regression tests
+Together, these ensure correctness, robustness, API stability, and compatibility with the JSO ecosystem.
+
+The tests cover, for all solvers:
+- Allocation behaviour of the implemented solvers (to detect unnecessary allocations and regressions);
+- The callback parameter and how it is invoked during the iterations;
+- Consistency of keyword arguments, to confirm that all solvers accept a common base set of options;
+- Proper error handling when an incompatible problem is passed;
+- Special options in solvers, such as `use_only_objgrad` for TRON;
+- Restart behaviour, i.e., re-solving a problem from a different initial guess while reusing previously allocated memory where possible;
+- GPU tests, to ensure that solvers work correctly with GPU-backed arrays when supported.
+
+These unit tests also act as regression tests: they serve to detect accidental behavioural changes when refactoring, updating dependencies, or extending solver capabilities.
+
+#### Convergence tests
+
+We use `SolverTest.jl` to run solvers on a collection of small analytic academic problems (unconstrained and bound-constrained) with different precision settings.
+These are black-box tests focused solely on:
+- reaching a first-order stationary point,
+- producing a valid solver status,
+- and respecting tolerances on optimality and feasibility.
+
+They do not enforce a specific iteration path or number of iterations; they only ensure that solver outputs remain correct and stable across releases.
 
 ## Dependencies
 
