### 2025-12-10 0cf5c5a
Remove tests that are duplicate from SolverTest
