+++ Project.toml
@@ -25,7 +25,6 @@ SolverTools = "0.10"
 julia = "1.10"
 
 [extras]
-ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 NLPModelsTest = "7998695d-6960-4d3a-85c4-e1bceb8cd856"
 SolverTest = "4343dc35-3317-4c6e-8877-f0cc8502c90e"
@@ -33,4 +32,4 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["ADNLPModels", "CUDA", "NLPModelsTest", "SolverTest", "SparseArrays", "Test"]
+test = ["CUDA", "NLPModelsTest", "SolverTest", "SparseArrays", "Test"]
