### 2025-12-12 ff6cd7b
Update R2N.jl
Remove the bug npcCount = 0 too early
