### 2025-12-12 42312cb
Switch NPC handler to Goldstein line search by default
Replaces the default non-positive curvature (NPC) handler from Armijo to Goldstein line search throughout the R2N solver. Updates documentation, keyword defaults, and implements the Goldstein line search logic for NPC cases, including bracketing and step size selection. Also includes minor optimizations and bug fixes, such as resetting counters and avoiding redundant objective evaluations.
