+++ src/R2N.jl
@@ -195,9 +195,9 @@ function HSLDirectSolver(nlp::AbstractNLPModel{T, V}, hsl_constructor) where {T,
   return HSLDirectSolver{T, typeof(hsl_obj)}(hsl_obj, rows, cols, vals, n, nnzh, work)
 end
 
-const npc_handler_allowed = [:armijo, :sigma, :prev, :cp]
+const npc_handler_allowed = [:gs, :sigma, :prev, :cp]
 
-const R2N_allowed_subsolvers = [:cg, :cr, :minres, :minres_qlp, :shifted_lbfgs, :ma97, :ma57] #TODO Prof. Orban check if I can add more
+const R2N_allowed_subsolvers = [:cg, :cr, :minres, :minres_qlp, :shifted_lbfgs, :ma97, :ma57]
 
 """
     R2N(nlp; kwargs...)
@@ -211,8 +211,6 @@ For advanced usage, first define a `R2NSolver` to preallocate the memory used in
 
 # Arguments
 - `nlp::AbstractNLPModel{T, V}` is the model to solve, see `NLPModels.jl`.
-
-
 
 # Keyword arguments
 - `params::R2NParameterSet = R2NParameterSet(nlp)`: algorithm parameters, see [`R2NParameterSet`](@ref).
@@ -236,8 +234,8 @@ For advanced usage, first define a `R2NSolver` to preallocate the memory used in
 - `subsolver::Symbol  = :cg`: the subsolver to solve the shifted system. The `MinresWorkspace` which solves the shifted linear system exactly at each iteration. Using the exact solver is only possible if `nlp` is an `LBFGSModel`. See `JSOSolvers.R2N_allowed_subsolvers` for a list of available subsolvers.
 - `subsolver_verbose::Int = 0`: if > 0, display iteration information every `subsolver_verbose` iteration of the subsolver if KrylovWorkspace type is selected.
 - `scp_flag::Bool = true`: if true, we compare the norm of the calculate step with `θ2 * norm(scp)`, each iteration, selecting the smaller step.
-- `npc_handler::Symbol = :armijo`: the non_positive_curve handling strategy.
-  - `:armijo`: uses the Armijo rule to handle non-positive curvature.
+- `npc_handler::Symbol = :gs`: the non_positive_curve handling strategy.
+  - `:gs`: uses the Goldstein conditions rule to handle non-positive curvature.
   - `:sigma`: increase the regularization parameter σ.
   - `:prev`: if subsolver return after first iteration, increase the sigma, but if subsolver return after second iteration, set s_k = s_k^(t-1).
   - `:cp`: set s_k to Cauchy point.
@@ -252,7 +250,7 @@ The value returned is a `GenericExecutionStats`, see `SolverCore.jl`.
 $(Callback_docstring)
 
 # Examples
-```jldoctest
+```jldoctest; output = false
 using JSOSolvers, ADNLPModels
 nlp = ADNLPModel(x -> sum(x.^2), ones(3))
 stats = R2N(nlp)
@@ -262,7 +260,7 @@ stats = R2N(nlp)
 "Execution stats: first-order stationary"
 ```
 
-```jldoctest
+```jldoctest; output = false
 using JSOSolvers, ADNLPModels
 nlp = ADNLPModel(x -> sum(x.^2), ones(3))
 solver = R2NSolver(nlp);
@@ -475,7 +473,7 @@ function SolverCore.solve!(
   max_iter::Int = typemax(Int),
   verbose::Int = 0,
   subsolver_verbose::Int = 0,
-  npc_handler::Symbol = :armijo,
+  npc_handler::Symbol = :gs,
   scp_flag::Bool = true,
 ) where {T, V}
[…]
