### 2025-12-14 JuliaSmoothOptimizers/JSOSolvers.jl#337 @dourouc05
Review from JOSS
Related to https://github.com/openjournals/joss-reviews/issues/9467#issuecomment-3649658463

I've got a few suggestions for this package overall. This is currently WIP!

For the docs:

- [ ] Formatting in https://jso.dev/JSOSolvers.jl/stable/internal/: `projected_line_search_ls!` and `cauchy_ls!` do not have the same formatting as the other functions here.
- [ ] User friendliness in https://jso.dev/JSOSolvers.jl/stable/solvers/: if you stumble upon the package, it's unlikely you will understand NLP/NLS.
- [ ] https://jso.dev/JSOSolvers.jl/stable/solvers/#JSOSolvers.lbfgs: maybe a link on `GenericExecutionStats`/`SolverCore.jl`?
- [ ] As you have callbacks for every solver, maybe the corresponding explanations could be made into their own section?
- [ ] Is there a link from the docs to the tutorials (https://jso.dev/tutorials/)?

Overall, the docs are quite good!

In the paper:

- [ ] "Julia’s native C++ and Fortran interoperability": I didn't know about C++, but the C part is quite good.
- [ ] Benchmarking: I would find it interesting if you could benchmark against the canonical implementations that you are referring to in the paper (like the original TRON code). I understand it might be quite a bit of work, hence feel free to ignore!

Code: WIP
