### 2025-12-15 9a1eb85
fix the cp and sigma case
