### 2025-12-17 b35dcf6
Remove tests that are duplicate from SolverTest - Fix docs (#336)
* Remove tests that are duplicate from SolverTest
