### 2025-12-17 b91ca7d
Remove tests that are duplicate from SolverTest - Fix docs (#336)
* Remove tests that are duplicate from SolverTest
