+++ test/test_solvers.jl
@@ -81,6 +81,3 @@ function tests()
 end
 
 tests()
-
-include("solvers/trunkls.jl")
-include("incompatible.jl")
