### 2025-12-17 c128df0
Add deps and tests comment in the documentation (#334)
* Add deps and tests comment

* Update docs/src/index.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* add-more-on-tests

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
