### 2025-12-17 0b09d4c
Fix sentence on interoperability
