### 2025-12-17 a32ca7b
Update src/tronls.jl
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
