### 2025-12-19 JuliaSmoothOptimizers/JSOSolvers.jl#337 @dourouc05
I ticked the boxes for the changes for which you have opened PRs that I have approved. (I hope that this doesn't make me an approver for this package :D!)

For the other changes, feel free to open other issues (for instance, referring to this one) as you see fit. For the code, I think you have issues that cover what I mean, which is enough.
