### 2025-12-26 274aab6
Fix sentence on interoperability (#340)
