+++ src/tron.jl
@@ -82,7 +82,7 @@ The keyword arguments may include
 The keyword arguments of `TronSolver` are passed to the [`TRONTrustRegion`](https://jso.dev/SolverTools.jl/stable/95-reference/#SolverTools.TRONTrustRegion) constructor.
 
 # Output
-The value returned is a `GenericExecutionStats`, see `SolverCore.jl`.
+The value returned is a [`GenericExecutionStats`](https://jso.dev/SolverCore.jl/stable/95-reference/#SolverCore.GenericExecutionStats), see `SolverCore.jl`.
 
 # Callback
 $(Callback_docstring)
@@ -498,7 +498,9 @@ function projected_line_search!(
   return s
 end
 
-"""`α, s = cauchy!(x, H, g, Δ, ℓ, u, s, Hs; μ₀ = 1e-2, μ₁ = 1.0, σ=10.0)`
+"""
+
+    α, s = cauchy!(x, H, g, Δ, ℓ, u, s, Hs; μ₀ = 1e-2, μ₁ = 1.0, σ = 10.0)
 
 Computes a Cauchy step `s = P(x - α g) - x` for
 
@@ -588,7 +590,7 @@ end
 
 Compute an approximate solution `d` for
 
-min q(d) = ¹/₂dᵀHs + dᵀg    s.t.    ℓ ≦ x + d ≦ u,  ‖d‖ ≦ Δ
+    min q(d) = ¹/₂dᵀHs + dᵀg    s.t.    ℓ ≦ x + d ≦ u,  ‖d‖ ≦ Δ
 
 starting from `s`.  The steps are computed using the conjugate gradient method
 projected on the active bounds.
