### 2025-12-26 a6d7d7d
Improve documentation and fix docstrings (#339)
* Improve doc

* Update src/tronls.jl

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
