+++ src/fomo.jl
@@ -157,21 +157,6 @@ For advanced usage, first define a `FomoSolver` to preallocate the memory used i
 
 The value returned is a [`GenericExecutionStats`](https://jso.dev/SolverCore.jl/stable/95-reference/#SolverCore.GenericExecutionStats), see `SolverCore.jl`.
 
-The callback is called at each iteration.
-The expected signature of the callback is `callback(nlp, solver, stats)`, and its output is ignored.
-Changing any of the input arguments will affect the subsequent iterations.
-In particular, setting `stats.status = :user || stats.stats = :unknown` will stop the algorithm.
-All relevant information should be available in `nlp` and `solver`.
-Notably, you can access, and modify, the following:
-- `solver.x`: current iterate;
-- `solver.gx`: current gradient;
-- `stats`: structure holding the output of the algorithm (`GenericExecutionStats`), which contains, among other things:
-    - `stats.dual_feas`: norm of current gradient;
-    - `stats.iter`: current iteration counter;
-    - `stats.objective`: current objective function value;
-    - `stats.status`: current status of the algorithm. Should be `:unknown` unless the algorithm has attained a stopping criterion. Changing this to anything will stop the algorithm, but you should use `:user` to properly indicate the intention.
-    - `stats.elapsed_time`: elapsed time in seconds.
-
 # Examples
 
 ## `fomo`
@@ -301,10 +286,6 @@ For advanced usage, first define a `FomoSolver` to preallocate the memory used i
 # Output
 
 The value returned is a `GenericExecutionStats`, see `SolverCore.jl`.
-
-# Callback
-
-$(Callback_docstring)
 
 # Examples
 
