### 2025-12-29 f4491bc
Move callback section to doc (#341)
* Move callback section to doc

* fix
