### 2025-12-29 467ac3c
Move callback section to doc (#341)
* Move callback section to doc

* fix
