### 2026-01-07 JuliaSmoothOptimizers/JSOSolvers.jl#343 @nkrusch
JOSS paper review comments
I have the following comments about the paper - ref: [#9467](https://github.com/openjournals/joss-reviews/issues/9467).

Line numbers refer to [paper.md](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/274aab6a998a7e8206571b5553886e2cdeeceabb/paper/paper.md) (not rendered draft).

- [x] [L43](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/274aab6a998a7e8206571b5553886e2cdeeceabb/paper/paper.md?plain=1#L43C356-L43C383): missing space after `Percival.jl`, before bracket.

- [x] [L49](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/274aab6a998a7e8206571b5553886e2cdeeceabb/paper/paper.md?plain=1#L49): Using "latter" for more than two items; consider re-wording (it is confusing which solver latter refers to).

- [x] [L61](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/274aab6a998a7e8206571b5553886e2cdeeceabb/paper/paper.md?plain=1#L61): TRUNK-NLS first appears here (what is it?); consider introducing earlier (for example, around paragraph L48-53).

- [x] L77, 87, 95-98, 101: mixed case of IPOPT vs. Ipopt: I am confused by the different capitalization styles.
  If mixed case is necessary, it should at least be used consistently (say, to separate theory vs. implementation).
  E.g., At L77 "Solver like IPOPT" and L101 "Ipopt ... solver" presumably refer to the same solver?

- [x] [L87](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/274aab6a998a7e8206571b5553886e2cdeeceabb/paper/paper.md?plain=1#L87): Perhaps missing a noun here: _"Ipopt is widely used and efficient but requires explicit derivatives..."_.
   Compare: _"Ipopt is a widely used and efficient solver, but requires explicit derivatives..."_.

- [x] Benchmarking: what kind of machine was used here (architecture, GPU/CPU, memory, cores)?
     This can be in a footnote, but it should be somewhere, to make sense of the reported times.
