### 2026-01-07 JuliaSmoothOptimizers/JSOSolvers.jl#344 @nkrusch
JOSS paper bibliography review comments
Specifically about the [bibliography](https://github.com/JuliaSmoothOptimizers/JSOSolvers.jl/blob/274aab6a998a7e8206571b5553886e2cdeeceabb/paper/paper.bib) - ref: [#9467](https://github.com/openjournals/joss-reviews/issues/9467).

* For all of the following: The Zenodo entry has different title and author list (often partial).
  Since they are yours, can you make the Zenodo meta data and bibliography entries consistent?
    * [x] Antunes dos Santos, E., Migot, T., Orban, D., Soares Siqueira, A., & contributors. (2023). Percival.jl.
    * [x] Migot, T., & contributors. (2023). FletcherPenaltySolver.jl
    * [x] Migot, T., Orban, D., Soares Siqueira, A., & contributors. (2024). JuliaSmoothOptimizers
    * [x] Orban, D., Soares Siqueira, A., & contributors. (2023a). NLPModels.jl
    * [x] Orban, D., Soares Siqueira, A., & contributors. (2023b). SolverBenchmark.jl
          (but DOI links to CaNNOLeS.jl)

The PR #345 includes style edits and fixes:

* [x] Double-check entry NLPModels.jl: DOI was referencing Version v0.7.1, but bib entry says v0.20.0.
  fix: changed the DOI to refer to v. 0.20.0.

* [x] Double-check entry Fourer (2003): the DOI refers to an article from 1990.
   fix: changed the bib to match the DOI info (from 1990).

* [x] Double-check entry Shin et al.: DOI [10.2139/ssrn.4601442](https://doi.org/10.2139/ssrn.4601442) links to a 2023 preprint, but the paper was actually published in 2024 at [10.1016/j.epsr.2024.110651](https://doi.org/10.1016/j.epsr.2024.110651).
  fix: changed the DOI to the 2024 publication
