+++ paper/paper.bib
@@ -104,7 +104,7 @@ doi = {10.5281/zenodo.7764782}
   title={A {M}odeling {L}anguage for {M}athematical {P}rogramming},
   author={Fourer, Robert and Gay, David M. and Kernighan, Brian W.},
   year={1990},
-  publisher={Management Science},
+  journal={Management Science},
   isbn={0-534-38809-4},
   doi = {10.1287/mnsc.36.5.519},
   volume = {36},
