### 2026-01-09 a570776
bibliography edits (#345)
* bibliography edits

* change publisher to journal
