+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
+  title={Julia: A {F}resh {A}pproach to {N}umerical {C}omputing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
   journal={SIAM Review},
   volume={59},
@@ -19,12 +19,12 @@
   pages={110651},
   year={2024},
   publisher={Elsevier},
-  doi={10.2139/ssrn.4601442}
+  doi={10.1016/j.epsr.2024.110651}
 }
 
 # Citation for R2, FOMO is an extension.
 @article{aravkin2022proximal,
-  title={A proximal quasi-Newton trust-region method for nonsmooth regularized optimization},
+  title={A {P}roximal {Q}uasi-{N}ewton {T}rust-{R}egion {M}ethod for {N}onsmooth {R}egularized {O}ptimization},
   author={Aravkin, Aleksandr Y and Baraldi, Robert and Orban, Dominique},
   journal={SIAM Journal on Optimization},
   volume={32},
@@ -86,12 +86,12 @@ title = {{{NLPModels.jl: Data Structures for Optimization Models}}},
 url = {https://github.com/JuliaSmoothOptimizers/NLPModels.jl},
 version = {0.20.0},
 year = {2023},
-doi = {10.5281/zenodo.2558627}
+doi = {10.5281/zenodo.7764782}
 }
 
 @article{jump,
   author = {Iain Dunning and Joey Huchette and Miles Lubin},
-  title = {{JuMP}: A Modeling Language for Mathematical Optimization},
+  title = {{JuMP}: A {M}odeling {L}anguage for {M}athematical {O}ptimization},
   journal = {SIAM Review},
   volume = {59},
   number = {2},
@@ -100,19 +100,21 @@ doi = {10.5281/zenodo.2558627}
   doi = {10.1137/15M1020575},
 }
 
-@book{fourer2003ampl,
-  title={AMPL. A modeling language for mathematical programming},
+@article{fourer2003ampl,
+  title={A {M}odeling {L}anguage for {M}athematical {P}rogramming},
   author={Fourer, Robert and Gay, David M. and Kernighan, Brian W.},
-  year={2003},
-  publisher={2nd ed., Brooks/Cole, Pacific Grove, CA},
+  year={1990},
+  journal={Management Science},
   isbn={0-534-38809-4},
   doi = {10.1287/mnsc.36.5.519},
+  volume = {36},
+  number = {5},
+  pages = {519--554}
 }
 
 @article{cutest,
   author = {Gould, Nicholas I. and Orban, Dominique and Toint, Philippe L.},
-  title = {{CUTEst}: A Constrained and Unconstrained Testing Environment with Safe
-           Threads for Mathematical Optimization},
+  title = {{CUTEst}: A {C}onstrained and {U}nconstrained {T}esting {E}nvironment with safe threads for mathematical optimization},
   journal = {Computational Optimization and Applications},
   volume = {60},
   number = {3},
@@ -122,7 +124,7 @@ doi = {10.5281/zenodo.2558627}
 }
 
 @article{tron,
-  title = {Newton’s Method for Large Bound-Constrained Optimization Problems},
+  title = {Newton’s {M}ethod for {L}arge {B}ound-{C}onstrained {O}ptimization {P}roblems},
   volume = {9},
   ISSN = {1095-7189},
   url = {http://dx.doi.org/10.1137/S1052623498345075},
@@ -143,7 +145,7 @@ journal = {Journal of Open Source Software},
 month = sep,
 number = {89},
 pages = {5187},
[…]
