### 2026-01-11 26b8b83
Update metadata and summary in paper.md
