+++ src/R2N.jl
@@ -1,0 +1,1053 @@
+export R2N, R2NSolver, R2NParameterSet
+export ShiftedLBFGSSolver, HSLDirectSolver
+export ShiftedOperator
+
+using LinearOperators, LinearAlgebra
+using SparseArrays
+using HSL
+
+#TODO move to LinearOperators
+# Define a new mutable operator for A = H + σI
+mutable struct ShiftedOperator{T, V, OpH <: Union{AbstractLinearOperator{T}, AbstractMatrix{T}}} <:
+               AbstractLinearOperator{T}
+  H::OpH
+  σ::T
+  n::Int
+  symmetric::Bool
+  hermitian::Bool
+end
+
+function ShiftedOperator(
+  H::OpH,
+) where {T, OpH <: Union{AbstractLinearOperator{T}, AbstractMatrix{T}}}
+  is_sym = isa(H, AbstractLinearOperator) ? H.symmetric : issymmetric(H)
+  is_herm = isa(H, AbstractLinearOperator) ? H.hermitian : ishermitian(H)
+  return ShiftedOperator{T, Vector{T}, OpH}(H, zero(T), size(H, 1), is_sym, is_herm)
+end
+
+# Define required properties for AbstractLinearOperator
+Base.size(A::ShiftedOperator) = (A.n, A.n)
+LinearAlgebra.isreal(A::ShiftedOperator{T}) where {T <: Real} = true
+LinearOperators.issymmetric(A::ShiftedOperator) = A.symmetric
+LinearOperators.ishermitian(A::ShiftedOperator) = A.hermitian
+
+# Define the core multiplication rules: y = (H + σI)x
+function LinearAlgebra.mul!(
+  y::AbstractVecOrMat,
+  A::ShiftedOperator{T, V, OpH},
+  x::AbstractVecOrMat{T},
+) where {T, V, OpH}
+  mul!(y, A.H, x)
+  LinearAlgebra.axpy!(A.σ, x, y)
+  return y
+end
+
+"""
+    R2NParameterSet([T=Float64]; θ1, θ2, η1, η2, γ1, γ2, γ3, σmin, non_mono_size)
+Parameter set for the R2N solver. Controls algorithmic tolerances and step acceptance.
+# Keyword Arguments
+- `θ1 = T(0.5)`: Cauchy step parameter (0 < θ1 < 1).
+- `θ2 = eps(T)^(-1)`: Cauchy step parameter.
+- `η1 = eps(T)^(1/4)`: Step acceptance parameter (0 < η1 ≤ η2 < 1).
+- `η2 = T(0.95)`: Step acceptance parameter (0 < η1 ≤ η2 < 1).
+- `γ1 = T(1.5)`: Regularization update parameter (1 < γ1 ≤ γ2).
+- `γ2 = T(2.5)`: Regularization update parameter (γ1 ≤ γ2).
+- `γ3 = T(0.5)`: Regularization update parameter (0 < γ3 ≤ 1).
+- `δ1 = T(0.5)`: Cauchy point calculation parameter.
+- `σmin = eps(T)`: Minimum step parameter. #TODO too small I need it to be 1
+- `non_mono_size = 1`: the size of the non-monotone behaviour. If > 1, the algorithm will use a non-monotone strategy to accept steps.
+"""
+struct R2NParameterSet{T} <: AbstractParameterSet
+  θ1::Parameter{T, RealInterval{T}}
+  θ2::Parameter{T, RealInterval{T}}
+  η1::Parameter{T, RealInterval{T}}
+  η2::Parameter{T, RealInterval{T}}
+  γ1::Parameter{T, RealInterval{T}}
+  γ2::Parameter{T, RealInterval{T}}
[…]
