+++ docs/src/solvers.md
@@ -7,11 +7,13 @@
 - [`trunk`](@ref)
 - [`R2`](@ref)
 - [`fomo`](@ref)
+- [`R2N`](@ref)
+- [`R2NLS`](@ref)
 
 | Problem type          | Solvers  |
 | --------------------- | -------- |
-| Unconstrained Nonlinear Optimization Problem     | [`lbfgs`](@ref), [`tron`](@ref), [`trunk`](@ref), [`R2`](@ref), [`fomo`](@ref)|
-| Unconstrained Nonlinear Least Squares     | [`trunk`](@ref), [`tron`](@ref) |
+| Unconstrained Nonlinear Optimization Problem     | [`lbfgs`](@ref), [`tron`](@ref), [`trunk`](@ref), [`R2`](@ref), [`fomo`](@ref), ['R2N'](@ref)|
+| Unconstrained Nonlinear Least Squares     | [`trunk`](@ref), [`tron`](@ref), ['R2NLS'](@ref)|
 | Bound-constrained Nonlinear Optimization Problem  | [`tron`](@ref) |
 | Bound-constrained Nonlinear Least Squares | [`tron`](@ref) |
 
@@ -23,4 +25,6 @@ tron
 trunk
 R2
 fomo
+R2N
+R2NLS
 ```

[…]
