### 2026-01-13 cbef2c2
Update paper.bib
