+++ paper/paper.bib
@@ -36,25 +36,24 @@
 }
 
 @software{Percival_jl,
-author = {Antunes dos Santos, Egmara and Migot, Tangi and Orban, Dominique and Soares Siqueira, Abel and contributors},
-license = {MPL-2.0},
-month = jul,
-title = {{{Percival.jl: an augmented Lagrangian method}}},
-url = {https://github.com/JuliaSmoothOptimizers/Percival.jl},
-version = {0.7.0},
-year = {2023},
-doi = {10.5281/zenodo.3969045}
+  doi = {10.5281/ZENODO.3969045},
+  url = {https://github.com/JuliaSmoothOptimizers/Percival.jl},
+  author = {Antunes dos Santos,  Egmara and Migot,  Tangi and Orban,  Dominique and Soares Siqueira,  Abel and contributors,  },
+  title = {Percival.jl: an augmented {L}agrangian method},
+  version = {0.7.6},
+  year = {2026},
+  license = {MPL-2.0}
 }
 
 @software{FletcherPenaltySolver_jl,
-author = {Migot, Tangi and contributors},
-license = {MPL-2.0},
-month = jul,
-title = {{{FletcherPenaltySolver.jl: Fletcher's penalty method for nonlinear optimization models}}},
-url = {https://github.com/JuliaSmoothOptimizers/FletcherPenaltySolver.jl},
-version = {0.2.4},
-year = {2023},
-doi = {10.5281/zenodo.7153564}
+  license = {MPL-2.0},
+  url = {https://github.com/JuliaSmoothOptimizers/FletcherPenaltySolver.jl},
+  version = {0.3.0},
+  doi = {10.5281/ZENODO.7153563},
+  author = {Migot,  Tangi and Orban,  Dominique and Soares Siqueira,  Abel and contributors,  },
+  keywords = {Nonlinear Optimization,  Julia,  Nonlinear Programming},
+  title = {{FletcherPenaltySolver.jl}: {F}letcher's penalty method for nonlinear optimization models},
+  year = {2025},
 }
 
 @software{OptimizationProblems_jl,
@@ -68,25 +67,15 @@ year = {2023},
 doi = {10.5281/zenodo.3672094}
 }
 
-@software{The_JuliaSmoothOptimizers_Ecosystem,
-author = {Migot, Tangi and Orban, Dominique and Soares Siqueira, Abel and contributors},
-license = {MPL-2.0},
-month = feb,
-title = {{The JuliaSmoothOptimizers Ecosystem for Numerical Linear Algebra and Optimization in Julia}},
-doi = {10.5281/zenodo.2655082},
-url = {https://jso.dev},
-year = {2024}
-}
-
 @software{NLPModels_jl,
-author = {Orban, Dominique and Soares Siqueira, Abel and contributors},
-license = {MPL-2.0},
-month = mar,
-title = {{{NLPModels.jl: Data Structures for Optimization Models}}},
-url = {https://github.com/JuliaSmoothOptimizers/NLPModels.jl},
-version = {0.20.0},
-year = {2023},
-doi = {10.5281/zenodo.7764782}
+  license = {MPL-2.0},
+  url = {https://github.com/JuliaSmoothOptimizers/NLPModels.jl},
+  version = {0.21.7},
+  doi = {10.5281/ZENODO.2558626},
+  author = {Orban,  Dominique and Soares Siqueira,  Abel and contributors,  },
+  keywords = {Nonlinear Optimization,  Julia,  Nonlinear Programming},
+  title = {{NLPModels.jl}: Data Structures for Optimization Models},
+  year = {2026},
 }
 
 @article{jump,
[…]
