### 2026-01-13 5b28538
Update metadata and summary in paper.md
