### 2026-01-13 a2053be
Update paper/paper.md
