### 2026-01-13 5c82934
Update paper/paper.md
