### 2026-01-13 a2c43a1
Update paper metadata and summary content
