### 2026-01-14 JuliaSmoothOptimizers/JSOSolvers.jl#343 @nkrusch
@tmigot by #346, these are good, thanks!
