### 2026-01-14 JuliaSmoothOptimizers/JSOSolvers.jl#344 @nkrusch
Checked against #346 - also good!
