### 2026-01-14 503516b
Fix JOSS paper review comments (#346)
* Update metadata and summary in paper.md

* Update paper.bib

* Update metadata and summary in paper.md

* Update paper/paper.md

* Update paper/paper.md

* Update paper metadata and summary content
