### 2026-01-14 #9467 @nkrusch
@tmigot: thanks for resolving the earlier comments.

As the next step, I will need you to create a new release and archive it.
Please complete the following steps.

1. Make a (new) tagged release of the software.
1. Archive the reviewed software in Zenodo, Figshare, or similar archival service.
1. Check the archival deposit has the correct metadata. This includes:
   - title (should match the paper)
   - author list (should match the paper)
   - authors' ORCID (should match the paper)
   - license (should match the repository)
   - release version tag should match the archive version name
      For example, git tag v1.0 would have archive name v1.0 (displays as "Version v1.0" on Zenodo).

1. List here, as a comment:
   - the version tag
   - the archive DOI

I can then move forward with recommending accept for the submission.
