### 2026-01-14 811c226
Version 0.14.4 (#348)
JOSS release for paper submission.

This release corresponds to the version reviewed in the JOSS submission.
