### 2026-01-14 95ea202
Remove Zenodo and update citation
