### 2026-01-14 #9467 @tmigot
Thanks @nkrusch

Here are the information:
the version tag: v0.14.5
the archive DOI: https://doi.org/10.5281/zenodo.18247251
