### 2026-01-14 67cdb5c
minor changes in paper.md
