### 2026-01-14 1425bf6
case changes in paper.bib
