+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{bezanson2017julia,
-  title={Julia: A {F}resh {A}pproach to {N}umerical {C}omputing},
+  title={Julia: A Fresh Approach to Numerical Computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
   journal={SIAM Review},
   volume={59},
@@ -24,7 +24,7 @@
 
 # Citation for R2, FOMO is an extension.
 @article{aravkin2022proximal,
-  title={A {P}roximal {Q}uasi-{N}ewton {T}rust-{R}egion {M}ethod for {N}onsmooth {R}egularized {O}ptimization},
+  title={A Proximal Quasi-{N}ewton Trust-Region Method for Nonsmooth Regularized Optimization},
   author={Aravkin, Aleksandr Y and Baraldi, Robert and Orban, Dominique},
   journal={SIAM Journal on Optimization},
   volume={32},
@@ -80,7 +80,7 @@ doi = {10.5281/zenodo.3672094}
 
 @article{jump,
   author = {Iain Dunning and Joey Huchette and Miles Lubin},
-  title = {{JuMP}: A {M}odeling {L}anguage for {M}athematical {O}ptimization},
+  title = {{JuMP}: A Modeling Language for Mathematical Optimization},
   journal = {SIAM Review},
   volume = {59},
   number = {2},
@@ -90,7 +90,7 @@ doi = {10.5281/zenodo.3672094}
 }
 
 @article{fourer2003ampl,
-  title={A {M}odeling {L}anguage for {M}athematical {P}rogramming},
+  title={A Modeling Language for Mathematical Programming},
   author={Fourer, Robert and Gay, David M. and Kernighan, Brian W.},
   year={1990},
   journal={Management Science},
@@ -113,7 +113,7 @@ doi = {10.5281/zenodo.3672094}
 }
 
 @article{tron,
-  title = {Newton’s {M}ethod for {L}arge {B}ound-{C}onstrained {O}ptimization {P}roblems},
+  title = {Newton’s Method for Large Bound-Constrained Optimization Problems},
   volume = {9},
   ISSN = {1095-7189},
   url = {http://dx.doi.org/10.1137/S1052623498345075},
@@ -177,7 +177,7 @@ year = {2022}
 }
 
 @article{lubin2015computing,
-  title={Computing in {O}perations {R}esearch {U}sing {J}ulia},
+  title={Computing in Operations Research Using {J}ulia},
   author={Lubin, Miles and Dunning, Iain},
   journal={INFORMS Journal on Computing},
   volume={27},
@@ -232,7 +232,7 @@ doi = {10.5281/zenodo.10434673}
 }
 
 @book{conn2000trust,
-  title={Trust {R}egion {M}ethods},
+  title={Trust Region Methods},
   author={Conn, Andrew R. and Gould, Nicholas I.M. and Toint, Philippe L.},
   year={2000},
   publisher={SIAM},
@@ -247,7 +247,7 @@ doi = {10.5281/zenodo.10434673}
 }
 
 @incollection{byrd2006k,
-  title={{\sc{Knitro}}: An {I}ntegrated {P}ackage for {N}onlinear {O}ptimization},
+  title={{\sc{Knitro}}: An Integrated Package for Nonlinear Optimization},
   author={Byrd, Richard H and Nocedal, Jorge and Waltz, Richard A},
   booktitle={Large-{S}cale {N}onlinear {O}ptimization},
   pages={35--59},
@@ -294,3 +294,4 @@ doi = {10.5281/zenodo.10434673}
   month = jul,
   pages = {4882}
 }
+
