### 2026-01-14 #9467 @danielskatz
Congratulations to @tmigot (Tangi Migot) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers, and we could do this without you
